feat: register participant in a event

+++ src/main/java/com/victormanduca/eventbooking/infra/repositories/IParticipantsRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.victormanduca.eventbooking.domain.entities.Participants;
 
-public interface IParticipantsRepository extends JpaRepository<Participants, Long> {
+public interface IParticipantsRepository extends JpaRepository<Participants, Integer> {
 
 }
